unify with TinyVec
