+++ turbopack/crates/turbo-tasks-auto-hash-map/Cargo.toml
@@ -14,3 +14,11 @@ hashbrown = { workspace = true, features = ["serde"]}
 rustc-hash = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 shrink-to-fit = { workspace = true, features = ["hashbrown"] }
+
+[dev-dependencies]
+criterion = { workspace = true }
+turbo-tasks-malloc = { workspace = true }
+
+[[bench]]
+name = "tiny_vec"
+harness = false
