+++ turbopack/crates/turbo-tasks/Cargo.toml
@@ -65,7 +65,3 @@ criterion = { workspace = true, features = ["async_tokio"] }
 [[bench]]
 name = "mod"
 harness = false
-
-[[bench]]
-name = "tiny_vec"
-harness = false
